Move non-command functions to a package
Let the cmd deal with color, not the pkg
