+++ pkg/gotestdox/testdata/fuzz/FuzzPrettify/5838cdfae7b16cde2707c04599b62223e7bade8dafdd8a1c53b8f881e8f79d99
