+++ pkg/gotestdox/testdata/fuzz/FuzzPrettify/d72586fff5c01205
